move the oauth stuff to its own hook
