+++ app/ui-react/syndesis/src/modules/connections/components/WithConnectorForm.tsx
@@ -11,7 +11,7 @@ import {
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import i18n from '../../../i18n';
-import { parseValidationResult } from './utils';
+import { parseValidationResult } from '../utils';
 
 export interface IWithConnectorFormChildrenProps {
   /**
