+++ app/ui-react/syndesis/src/modules/connections/utils.ts
